Test Peer model and fix address field
Since the address field was a subclass of `model.CharField`, the
`Peer.maxip()` method returned the highest IP in alphanumerical order,
which caused trouble in finding the highest ip from the numerical point
of view. Changing the field to a subclass of
`model.PositiveBigIntegerField` solves the ordering problem and the
`model.Max` function runs correctly, so the `Peer.maxip()` method does.